refactor(studio): extract DatabaseSizeWidget from database report page
Moves the database-size section (space used, provisioned disk size,
increase disk size button, large objects table, disk size modal, and
informational alert) out of the inline JSX in database.tsx into a new
co-located component at components/interfaces/Reports/DatabaseSizeWidget.tsx.
All data fetching and state remain in the page; only the JSX is extracted.
