+++ apps/studio/pages/project/[ref]/observability/database.tsx
@@ -2,26 +2,22 @@ import { PermissionAction } from '@supabase/shared-types/out/constants'
 import { useQueryClient } from '@tanstack/react-query'
 import { useFlag, useParams } from 'common'
 import dayjs from 'dayjs'
-import { ArrowRight, ExternalLink, RefreshCw } from 'lucide-react'
-import Link from 'next/link'
+import { ArrowRight, RefreshCw } from 'lucide-react'
 import { useEffect, useRef, useState } from 'react'
 import { toast } from 'sonner'
-import { Alert, AlertDescription, Button } from 'ui'
+import { Button } from 'ui'
 
+import { DatabaseSizeWidget } from '@/components/interfaces/Reports/DatabaseSizeWidget'
 import ReportHeader from '@/components/interfaces/Reports/ReportHeader'
 import ReportPadding from '@/components/interfaces/Reports/ReportPadding'
 import { REPORT_DATERANGE_HELPER_LABELS } from '@/components/interfaces/Reports/Reports.constants'
 import ReportStickyNav from '@/components/interfaces/Reports/ReportStickyNav'
-import ReportWidget from '@/components/interfaces/Reports/ReportWidget'
 import { ReportChartUpsell } from '@/components/interfaces/Reports/v2/ReportChartUpsell'
 import { POOLING_OPTIMIZATIONS } from '@/components/interfaces/Settings/Database/ConnectionPooling/ConnectionPooling.constants'
-import DiskSizeConfigurationModal from '@/components/interfaces/Settings/Database/DiskSizeConfigurationModal'
 import { LogsDatePicker } from '@/components/interfaces/Settings/Logs/Logs.DatePickers'
 import UpgradePrompt from '@/components/interfaces/Settings/Logs/UpgradePrompt'
 import DefaultLayout from '@/components/layouts/DefaultLayout'
 import ObservabilityLayout from '@/components/layouts/ObservabilityLayout/ObservabilityLayout'
-import Table from '@/components/to-be-cleaned/Table'
-import { ButtonTooltip } from '@/components/ui/ButtonTooltip'
 import type { MultiAttribute } from '@/components/ui/Charts/ComposedChart.utils'
 import { LazyComposedChartHandler } from '@/components/ui/Charts/ComposedChartHandler'
 import { ReportSettings } from '@/components/ui/Charts/ReportSettings'
@@ -41,8 +37,6 @@ import { useAsyncCheckPermissions } from '@/hooks/misc/useCheckPermissions'
 import { useRefreshHandler, useReportDateRange } from '@/hooks/misc/useReportDateRange'
 import { useSelectedOrganizationQuery } from '@/hooks/misc/useSelectedOrganization'
 import { useSelectedProjectQuery } from '@/hooks/misc/useSelectedProject'
-import { DOCS_URL } from '@/lib/constants'
-import { formatBytes } from '@/lib/helpers'
 import { useDatabaseSelectorStateSnapshot } from '@/state/database-selector'
 import { SHORTCUT_IDS } from '@/state/shortcuts/registry'
 import { useShortcut } from '@/state/shortcuts/useShortcut'
@@ -338,122 +332,20 @@ const DatabaseUsage = () => {
           />
         )}
       </ReportStickyNav>
-      <section id="database-size-report">
-        <ReportWidget
-          isLoading={isLoading}
-          params={params.largeObjects}
-          title="Database Size"
-          data={data.largeObjects || []}
-          queryType={'db'}
-          resolvedSql={largeObjectsSql}
-          renderer={(props) => {
-            return (
-              <div>
-                <div className="col-span-4 inline-grid grid-cols-12 gap-12 w-full mt-5">
-                  <div className="grid gap-2 col-span-4 xl:col-span-2">
-                    <h5>Space used</h5>
-                    <span className="text-lg">{formatBytes(databaseSizeBytes, 2, 'GB')}</span>
-                  </div>
-                  <div className="grid gap-2 col-span-4 xl:col-span-3">
-                    <h5>Provisioned disk size</h5>
-                    <span className="text-lg">{currentDiskSize} GB</span>
-                  </div>
-
-                  <div className="col-span-full lg:col-span-4 xl:col-span-7 lg:text-right">
-                    {project?.cloud_provider === 'AWS' ? (
-                      <Button asChild type="default">
-                        <Link href={`/project/${ref}/settings/compute-and-disk`}>
-                          Increase disk size
-                        </Link>
-                      </Button>
-                    ) : (
-                      <ButtonTooltip
-                        type="default"
-                        disabled={!canUpdateDiskSizeConfig}
-                        onClick={() => setshowIncreaseDiskSizeModal(true)}
-                        tooltip={{
-                          content: {
-                            side: 'bottom',
-                            text: !canUpdateDiskSizeConfig
-                              ? 'You need additional permissions to increase the disk size'
-                              : undefined,
-                          },
-                        }}
-                      >
-                        Increase disk size
-                      </ButtonTooltip>
-                    )}
-                  </div>
-                </div>
-
-                <h3 className="mt-8 text-sm">Large Objects</h3>
-                {!props.isLoading && props.data.length === 0 && <span>No large objects found</span>}
-                {!props.isLoading && props.data.length > 0 && (
-                  <Table
-                    className="space-y-3 mt-4"
-                    head={[
-                      <Table.th key="object" className="py-2">
-                        Object
-                      </Table.th>,
-                      <Table.th key="size" className="py-2">
-                        Size
-                      </Table.th>,
-                    ]}
-                    body={props.data?.map((object) => {
-                      const percentage = (
-                        ((object.table_size as number) / databaseSizeBytes) *
-                        100
-                      ).toFixed(2)
-
-                      return (
-                        <Table.tr key={`${object.schema_name}.${object.relname}`}>
-                          <Table.td>
-                            {object.schema_name}.{object.relname}
-                          </Table.td>
-                          <Table.td>
-                            {formatBytes(object.table_size)} ({percentage}%)
-                          </Table.td>
-                        </Table.tr>
-                      )
-                    })}
-                  />
-                )}
-              </div>
-            )
-          }}
-          append={() => (
-            <div className="px-6 pb-6">
-              <Alert variant="default" className="mt-4">
-                <AlertDescription>
-                  <div className="space-y-2">
-                    <p>
-                      New Supabase projects have a database size of ~40-60mb. This space includes
-                      pre-installed extensions, schemas, and default Postgres data. Additional
-                      database size is used when installing extensions, even if those extensions are
-                      inactive.
-                    </p>
-
-                    <Button asChild type="default" icon={<ExternalLink />}>
-                      <Link
-                        href={`${DOCS_URL}/guides/platform/database-size#disk-space-usage`}
-                        target="_blank"
-                        rel="noreferrer"
-                      >
-                        Read about database size
-                      </Link>
-                    </Button>
-                  </div>
-                </AlertDescription>
-              </Alert>
-            </div>
-          )}
-        />
-        <DiskSizeConfigurationModal
-          visible={showIncreaseDiskSizeModal}
-          loading={isUpdatingDiskSize}
-          hideModal={setshowIncreaseDiskSizeModal}
-        />
-      </section>
+      <DatabaseSizeWidget
+        isLoading={isLoading}
+        params={params.largeObjects}
+        data={data.largeObjects || []}
+        resolvedSql={largeObjectsSql}
+        databaseSizeBytes={databaseSizeBytes}
+        currentDiskSize={currentDiskSize}
+        projectRef={ref}
+        cloudProvider={project?.cloud_provider}
+        canUpdateDiskSizeConfig={canUpdateDiskSizeConfig}
+        showIncreaseDiskSizeModal={showIncreaseDiskSizeModal}
+        isUpdatingDiskSize={isUpdatingDiskSize}
+        setShowIncreaseDiskSizeModal={setshowIncreaseDiskSizeModal}
+      />
       <div className="py-8">
         <ObservabilityLink />
       </div>
